test: setup testing infrastructure and modernize repository structure
