+++ Hm.Logging.sln
@@ -1,34 +1,39 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
+# Visual Studio Version 18
+VisualStudioVersion = 18.6.11806.211 stable
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hm.Logging", "Hm.Logging\Hm.Logging.csproj", "{B1382F16-6688-47C2-A35C-EC48E8C28990}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hm.Logging", "src\Hm.Logging\Hm.Logging.csproj", "{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hm.Logging.Tests", "tests\Hm.Logging.Tests\Hm.Logging.Tests.csproj", "{C2F9AB64-0BB4-457B-822E-E38A40B404AF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{82681B27-E484-4E1B-A238-8ADC07F60602}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{9538172B-5549-4F3F-B09C-428AD6A10390}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|x64.Build.0 = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Debug|x86.Build.0 = Debug|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|x64.ActiveCfg = Release|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|x64.Build.0 = Release|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|x86.ActiveCfg = Release|Any CPU
-		{B1382F16-6688-47C2-A35C-EC48E8C28990}.Release|x86.Build.0 = Release|Any CPU
+		{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2F9AB64-0BB4-457B-822E-E38A40B404AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2F9AB64-0BB4-457B-822E-E38A40B404AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2F9AB64-0BB4-457B-822E-E38A40B404AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2F9AB64-0BB4-457B-822E-E38A40B404AF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{3CC0661F-A836-E019-EF1C-FEBA9F7E3AB9} = {82681B27-E484-4E1B-A238-8ADC07F60602}
+		{C2F9AB64-0BB4-457B-822E-E38A40B404AF} = {9538172B-5549-4F3F-B09C-428AD6A10390}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D2E8C582-9AEE-491C-80C8-7CB8ACBC4805}
 	EndGlobalSection
 EndGlobal
