+++ src/Hm.Logging/Hm.Logging.csproj
