+++ tests/.editorconfig
@@ -1,0 +1,4 @@
+[*.cs]
+
+dotnet_diagnostic.CA1707.severity = none
+dotnet_diagnostic.IDE0058.severity = none
